Addition of file for demonstrating overprint when source colors are ICC based.
This goes with bug 692433. If was found that AR does not do overprint simulation when the source
ICC profile does not match the destination profile.  To test this, I created this
test file that had 2 overlapping fills with CMYK values that were specified by
SWOP CMYK values.  If the simulation profile in AR matched this profile, then overprinting
was properly displayed by AR.  If not, then overprinting was not displayed.  This
makes sense as the overprinted inks really should be in the color space for the
target device.   Hence, if the source and destination profiles do not match we will
not be showing overprint.